+++ TentakelsAttacking2/Helper/public/HFleetResult.hpp
@@ -3,7 +3,7 @@
 // 23.05.2023
 //
 
-#include "LogicAlias.hpp"
+#include "HLogicAlias.hpp"
 #include <memory>
 
 struct FleetResult final {
