+++ TentakelsAttacking2/Helper/public/HGalaxy.h
@@ -6,7 +6,7 @@
 #pragma once
 #include <utility>
 #include <memory>
-#include "LogicAlias.hpp"
+#include "HLogicAlias.hpp"
 
 [[nodiscard]] std::pair<bool, SpaceObject_ty> TryGetTarget(
 	Fleet_ty_raw fleet, SpaceObject_ty const& target);
