+++ TentakelsAttacking2/UI/Elements/Galaxy/public/UIGalaxy.h
@@ -7,7 +7,7 @@
 #include "UIElement.hpp"
 #include "EventListener.hpp"
 #include "Focusable.h"
-#include "LogicAlias.hpp"
+#include "HLogicAlias.hpp"
 #include <functional>
 #include <memory>
 #include <vector>
